+++ meteoio/meteostats/libinterpol1D.h
@@ -41,8 +41,8 @@ class Interpol1D {
 		static std::vector<double> quantiles(const std::vector<double>& X, const std::vector<double>& quartiles);
 		static std::vector<double> derivative(const std::vector<double>& X, const std::vector<double>& Y);
 		static void sort(std::vector<double>& X, std::vector<double>& Y, const bool& keep_nodata=true);
-		static void equalBin(const size_t k, std::vector<double> &X, std::vector<double> &Y);
-		static void equalCountBin(const size_t k, std::vector<double> &X, std::vector<double> &Y);
+		static void equalBin(const unsigned int k, std::vector<double> &X, std::vector<double> &Y);
+		static void equalCountBin(const unsigned int k, std::vector<double> &X, std::vector<double> &Y);
 		static double weightedMean(const double& d1, const double& d2, const double& weight=1.);
 		static double weightedMean(const std::vector<double>& vecData, const std::vector<double>& weight);
 		static double arithmeticMean(const std::vector<double>& vecData);
